feat: add pages folder
