+++ book-api-react-app/next.config.mjs
@@ -1,7 +1,7 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
-    output: 'export',
-    distDir: './dist',
-  }
-   
-  export default nextConfig
+  output: "export",
+  distDir: "./dist",
+};
+
+export default nextConfig;
