+++ book-api-react-app/src/components/Gallery/Card/Card.tsx
@@ -1,5 +1,5 @@
 import { useDispatch, useSelector } from "react-redux";
-import { BookInfo } from "../../../api/books";
+import { BookInfo } from "../../../redux/books";
 import LinkButton from "../../LinkButton/LinkButton";
 import { addItem, removeItem } from "../../../redux/selectedItemsSlice";
 import { RootState } from "../../../redux/store";
