+++ book-api-react-app/src/components/Loader/Loader.tsx
@@ -1,5 +1,3 @@
-import "./Loader.scss";
-
 export function Loader() {
   return (
     <div className="loading">
