+++ book-api-react-app/src/index.scss
@@ -1,3 +1,7 @@
+@import "./components/Button/Button.module.scss";
+@import "./components/SearchBar/SearchBar.module.scss";
+@import "./pages/Main.module.scss";
+
 $colorWhite: #fff;
 
 :root {
