--- book-api-react-app/src/pages/about.tsx
@@ -1,3 +1,0 @@
-export default function About() {
-    return <>About</>
-  }
