+++ book-api-react-app/src/pages/index.tsx
@@ -1,0 +1,16 @@
+import SearchBar from "../components/SearchBar/SearchBar";
+import "./Main.module.scss";
+import { useTheme } from "../components/ThemeContext/useTheme";
+
+export default function Main() {
+  const { theme } = useTheme();
+
+  return (
+    <div
+      className={theme === "light" ? "main-page" : "main-page main-page-dark"}
+    >
+      <SearchBar />
+      {/* <Gallery /> */}
+    </div>
+  );
+}
