--- book-api-react-app/src/pages/page404.tsx
@@ -1,5 +1,0 @@
-const Page404 = () => {
-  return <div>404</div>;
-};
-
-export default Page404;
